MyFlaskApp/requirements.txt
Flask==2.0.1
requests==2.25.1
Pillow==8.2.0
